requirements.txt
alembic==1.16.4
aniso8601==10.0.1
attrs==25.3.0
blinker==1.6.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dnspython==2.7.0
email-validator==2.0.0
Flask==2.2.5
Flask-Dance==7.0.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
gunicorn==20.1.0

idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.2
jsmin==3.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
lesscpy==0.15.1
Mako==1.3.10
MarkupSafe==3.0.2
oauthlib==3.3.1
openai==1.101.0
pdf2image==1.17.0
Pillow==11.3.0
ply==3.11
pyodbc==5.2.0
python-dotenv==0.19.2
pytz==2025.2
rcssmin==1.2.1
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.0
setuptools==80.9.0
SQLAlchemy==2.0.21
typing_extensions==4.14.1
urllib3==2.5.0
URLObject==3.0.0
Werkzeug==2.3.7
WTForms==3.0.1
xxhash==3.5.0
